Update: storing assets in tmaze/assets, where we have raw assets and ones for dist/.
