+++ tmaze/src/sound/track.rs
@@ -5,10 +5,30 @@ use cmaze::gameboard::Maze;
 use rand::{seq::SliceRandom, thread_rng};
 
 mod assets_sounds {
-    pub const MUSIC_EASY: &[u8] = include_bytes!(concat!(env!("OUT_DIR"), "/audio/", "music_easy-level.", env!("AUDIO_EXT")));
-    pub const MUSIC_MEDIUM: &[u8] = include_bytes!(concat!(env!("OUT_DIR"), "/audio/", "music_medium-level.", env!("AUDIO_EXT")));
-    pub const MUSIC_HARD: &[u8] = include_bytes!(concat!(env!("OUT_DIR"), "/audio/", "music_hard-level.", env!("AUDIO_EXT")));
-    pub const MUSIC_MENU: &[u8] = include_bytes!(concat!(env!("OUT_DIR"), "/audio/", "music_menu.", env!("AUDIO_EXT")));
+    pub const MUSIC_EASY: &[u8] = include_bytes!(concat!(
+        env!("CARGO_MANIFEST_DIR",),
+        "/assets/dist/audio/",
+        "music_easy-level.",
+        "mp3",
+    ));
+    pub const MUSIC_MEDIUM: &[u8] = include_bytes!(concat!(
+        env!("CARGO_MANIFEST_DIR",),
+        "/assets/dist/audio/",
+        "music_medium-level.",
+        "mp3",
+    ));
+    pub const MUSIC_HARD: &[u8] = include_bytes!(concat!(
+        env!("CARGO_MANIFEST_DIR",),
+        "/assets/dist/audio/",
+        "music_hard-level.",
+        "mp3",
+    ));
+    pub const MUSIC_MENU: &[u8] = include_bytes!(concat!(
+        env!("CARGO_MANIFEST_DIR",),
+        "/assets/dist/audio/",
+        "music_menu.",
+        "mp3",
+    ));
 }
 
 pub type Track = Box<dyn rodio::Source<Item = i16> + Send>;
